스트리밍 서비스 구현

+++ video_api/src/main/java/com/sparta/videoapi/repository/AdRepository.java
@@ -3,10 +3,5 @@ package com.sparta.videoapi.repository;
 import com.sparta.videoapi.entity.Ad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
-
-public interface AdRepository extends JpaRepository<Ad, Long> {
-//    List<Ad> findByVideoId(Long videoId);
-
+public interface AdRepository extends JpaRepository<Ad, Integer> {
 }
